refactor: 변경된 import 적용
