+++ src/main/java/com/mobble/mobbleserver/domain/clubMember/ClubMember.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.clubMember;
 
-import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
+import com.mobble.mobbleserver.domain.common.entity.BaseEntity;
 import com.mobble.mobbleserver.domain.article.ArticleType;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
