+++ src/main/java/com/mobble/mobbleserver/domain/notification/device/DeviceToken.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.notification.device;
 
-import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
+import com.mobble.mobbleserver.domain.common.entity.BaseEntity;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
